Organize testserver to enable running test scenarios
